Move upload command into own file
