+++ cli.js
@@ -476,33 +476,7 @@ yargs
   .command(require('./commands/privatekey'))
   .command(require('./commands/login'))
   .command(require('./commands/wallet'))
-  .command('upload [input]', 'upload a package without registry', yargs => {
-    yargs
-      .positional('input', {
-        describe: '.wbn package to upload',
-        // default: 5000
-      });
-  }, async argv => {
-    handled = true;
-
-    if (typeof argv.input !== 'string') {
-      argv.input = 'a.wbn';
-    }
-
-    const dataArrayBuffer = fs.readFileSync(argv.input);
-    const o = await uploadPackage(dataArrayBuffer, argv.input);
-    const {metadata, metadataHash} = o;
-    console.log('Name:', metadata.name);
-    console.log('Description:', metadata.description);
-    if (metadata.icons.length > 0) {
-      console.log('Icons:');
-      for (const o of metadata.icons) {
-        console.log(`  ${apiHost}/${o.hash} ${o.type}`);
-      }
-    }
-    console.log('Data:', `${apiHost}/${metadata.dataHash}.wbn`);
-    console.log('Metadata:', `${apiHost}/${metadataHash}.json`);
-  })
+  .command(require('./commands/upload'))
   .command('publish [input]', 'publish a package to ipfs', yargs => {
     yargs
       .positional('input', {
